login works with modal
login and protected pages are working. the main problem is that the components are showing up a tabs, which is definetly an issue with either the layout.tsx files, or how we are building our components. I took out the home.tsx code and it reduced the tabs by one, so maybe we need to write deeper classes or figure out how to only make 1 tsx file...... idrk